Extract EventCard to a separate file
